complete delete method

+++ src/TaskList.java
@@ -2,7 +2,7 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
-public class TaskList{
+public class TaskList {
     List<Task> taskList;
 
     public TaskList() {
@@ -15,8 +15,14 @@ public class TaskList{
         Database.insertTask(task);
     }
 
+    // Remove a task from list and db through task_id
     public void removeTask(int task_id) {
-        taskList.remove(task_id); //TODO
+        // search index of the task in arrayList
+        int index = searchIndex(task_id);
+        if (index < 0 || index >= size())
+            throw new IllegalArgumentException();
+
+        taskList.remove(index);
 
         Database.deleteTask(task_id);
     }
@@ -37,10 +43,30 @@ public class TaskList{
         return taskList.get(index);
     }
 
-    public void displayAll(){
+    public void displayAll() {
         int size = taskList.size();
         for (Task task : taskList) {
             System.out.println(task.toString());
         }
+    }
+
+    // Return index through binary search of task_id in list, if not found return -1
+    private int searchIndex(int task_id) {
+        int l = 0;
+        int r = size();
+
+        while (l <= r) {
+            int mid = l + (r - l) / 2;
+            int id = taskList.get(mid).getTask_id();
+            if (id == task_id)
+                return mid;
+
+            if (id > task_id)
+                r = mid - 1;
+            else
+                l = mid + 1;
+        }
+
+        return -1;
     }
 }
